cleanup: zero out seslib/__init__.py
Move the contents to more descriptively named Python files.
